refactor: create a renku.cli._format package
